pyproject.toml
[tool.poetry]
name = "work"
version = "0.1.0"
description = ""
authors = ["VictoriaS <viccisviri@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
python-decouple = "^3.5"
psycopg2-binary = "^2.9.2"
djangorestframework = "^3.13.1"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
pytest-factoryboy = "^2.1.0"
Faker = "^11.0.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
